[typos] rename tag in SQL physical query and update tests, locate sql in its appropriate directory
