+++ asistente_ladm_col/lib/queries/postgresql/basic_query.py
