+++ asistente_ladm_col/lib/queries/postgresql/economic_query.py
