+++ asistente_ladm_col/lib/queries/postgresql/legal_query.py
