+++ asistente_ladm_col/lib/queries/postgresql/physical_query.py
@@ -66,7 +66,7 @@ def get_igac_physical_query(schema, plot_t_id, parcel_fmi, parcel_number, previo
                                                                            'Tipo principal', col_fuenteespacial.tipo_principal,
                                                                            'Fecha de entrega', col_fuenteespacial.fecha_entrega,
                                                                            'Fecha de grabación', col_fuenteespacial.fecha_grabacion,
-                                                                           'Enlace fuente espacial', extarchivo.datos))
+                                                                           'Archivo fuente', extarchivo.datos))
             ORDER BY col_fuenteespacial.t_id) FILTER(WHERE ueFuente.pfuente IS NOT NULL) AS col_fuenteespacial
         FROM {schema}.uefuente LEFT JOIN {schema}.col_fuenteespacial ON uefuente.pfuente = col_fuenteespacial.t_id
         LEFT JOIN {schema}.extarchivo ON extarchivo.col_fuenteespacial_ext_archivo_id = col_fuenteespacial.t_id
@@ -123,7 +123,7 @@ def get_igac_physical_query(schema, plot_t_id, parcel_fmi, parcel_number, previo
                                                                            'Tipo principal', col_fuenteespacial.tipo_principal,
                                                                            'Fecha de entrega', col_fuenteespacial.fecha_entrega,
                                                                            'Fecha de grabación', col_fuenteespacial.fecha_grabacion,
-                                                                           'Enlace fuente espacial', extarchivo.datos))														   
+                                                                           'Archivo fuente', extarchivo.datos))														   
             ORDER BY col_fuenteespacial.t_id) FILTER(WHERE ueFuente.pfuente IS NOT NULL) AS col_fuenteespacial
         FROM {schema}.uefuente LEFT JOIN {schema}.col_fuenteespacial ON uefuente.pfuente = col_fuenteespacial.t_id
         LEFT JOIN {schema}.extarchivo ON extarchivo.col_fuenteespacial_ext_archivo_id = col_fuenteespacial.t_id
@@ -191,7 +191,7 @@ def get_igac_physical_query(schema, plot_t_id, parcel_fmi, parcel_number, previo
                                                                            'Tipo principal', col_fuenteespacial.tipo_principal,
                                                                            'Fecha de entrega', col_fuenteespacial.fecha_entrega,
                                                                            'Fecha de grabación', col_fuenteespacial.fecha_grabacion,
-                                                                           'Enlace fuente espacial', extarchivo.datos))														   
+                                                                           'Archivo fuente', extarchivo.datos))														   
             ORDER BY col_fuenteespacial.t_id) FILTER(WHERE ueFuente.pfuente IS NOT NULL) AS col_fuenteespacial
         FROM {schema}.uefuente LEFT JOIN {schema}.col_fuenteespacial ON uefuente.pfuente = col_fuenteespacial.t_id
         LEFT JOIN {schema}.extarchivo ON extarchivo.col_fuenteespacial_ext_archivo_id = col_fuenteespacial.t_id
