+++ asistente_ladm_col/lib/queries/postgresql/property_record_card_query.py
